+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Checkout client
+- Order payment approved event handler
+
 ## [0.1.0] - 2021-12-28
 
 ### Added